// src/api-client/models/create-employee-dto.ts
/* tslint:disable */

/**
 * API Documentation
 * API - IT Asset Tracker
 *
 * The version of the OpenAPI document: 0.0.1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface CreateEmployeeDto
 */
export interface CreateEmployeeDto {
  /**
   * Employee full name
   * @type {string}
   * @memberof CreateEmployeeDto
   */
  fullName: string;
  /**
   * Unique employee code (exactly 10 characters)
   * @type {string}
   * @memberof CreateEmployeeDto
   */
  employeeCode: string;
  /**
   * Employee email address
   * @type {string}
   * @memberof CreateEmployeeDto
   */
  email: string;
  /**
   * Employee password
   * @type {string}
   * @memberof CreateEmployeeDto
   */
  password: string;
  /**
   * Role identifier (foreign key)
   * @type {number}
   * @memberof CreateEmployeeDto
   */
  roleId: number;
}