// src/api-client/models/update-employee-dto.ts
/* tslint:disable */

/**
 * API Documentation
 * API - IT Asset Tracker
 *
 * The version of the OpenAPI document: 0.0.1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface UpdateEmployeeDto
 */
export interface UpdateEmployeeDto {
  /**
   * Employee full name
   * @type {string}
   * @memberof UpdateEmployeeDto
   */
  fullName?: string;
  /**
   * Unique employee code (exactly 10 characters)
   * @type {string}
   * @memberof UpdateEmployeeDto
   */
  employeeCode?: string;
  /**
   * Employee email address
   * @type {string}
   * @memberof UpdateEmployeeDto
   */
  email?: string;
  /**
   * Employee password
   * @type {string}
   * @memberof UpdateEmployeeDto
   */
  password?: string;
  /**
   * Role identifier (foreign key)
   * @type {number}
   * @memberof UpdateEmployeeDto
   */
  roleId?: number;
}